Separate roles and role permissions
