+++ shared_model/backend/protobuf/queries/proto_query.hpp
@@ -28,12 +28,13 @@
 #include "utils/variant_deserializer.hpp"
 
 #include "backend/protobuf/queries/proto_get_account.hpp"
+#include "backend/protobuf/queries/proto_get_account_asset_transactions.hpp"
 #include "backend/protobuf/queries/proto_get_account_assets.hpp"
+#include "backend/protobuf/queries/proto_get_account_transactions.hpp"
 #include "backend/protobuf/queries/proto_get_asset_info.hpp"
+#include "backend/protobuf/queries/proto_get_role_permissions.hpp"
 #include "backend/protobuf/queries/proto_get_roles.hpp"
 #include "backend/protobuf/queries/proto_get_signatories.hpp"
-#include "backend/protobuf/queries/proto_get_account_transactions.hpp"
-#include "backend/protobuf/queries/proto_get_account_asset_transactions.hpp"
 
 template <typename... T, typename Archive>
 auto load_query(Archive &&ar) {
