+++ shared_model/interfaces/queries/query.hpp
@@ -24,10 +24,11 @@
 #include "interfaces/polymorphic_wrapper.hpp"
 #include "interfaces/primitive.hpp"
 #include "interfaces/queries/get_account.hpp"
-#include "interfaces/queries/get_account_assets.hpp"
 #include "interfaces/queries/get_account_asset_transactions.hpp"
+#include "interfaces/queries/get_account_assets.hpp"
 #include "interfaces/queries/get_account_transactions.hpp"
 #include "interfaces/queries/get_asset_info.hpp"
+#include "interfaces/queries/get_role_permissions.hpp"
 #include "interfaces/queries/get_roles.hpp"
 #include "interfaces/queries/get_signatories.hpp"
 #include "interfaces/signable.hpp"
